Add disable logon trigger in case of problem with login trigger
If an error is caught, disable trigger otherwise if it is permanent, login will be denied to everybody.

In last commit, this was a reason why no direct reference  to FormatRunTimeMsg and FormatCurrentMsg is done. This could break the logon trigger if AuditReq database becomes unavailable. Since there is now very little change to break the trigger, I still add the disable trigger in case of error.